Updated note re durable use
Previously, the note warned against using Transit for durable applications. That note has been updated to instead note that it is the user's responsibility for readers and writers to either be compatible or to handle migration of durable data. The Transit format has thus far had only one version (0.8) and has not changed in several years.